kafka: rename kafka-destination.{c,h} to kafka-dest-driver.{c,h}
